+++ Makie/src/basic_recipes/tooltip.jl
@@ -1,8 +1,10 @@
 """
-    tooltip(position, string)
-    tooltip(x, y, string)
+Creates a tooltip pointing at a position displaying a given string.
 
-Creates a tooltip pointing at `position` displaying the given `string
+## Arguments
+
+* `position` Creates a tooltip at a given `position` of type `VecTypes` (`Point`, `Vec` or `Tuple`).
+* `x, y` Creates a tooltip at the given `x` and `y` coordinates fo type `<:Real`.
 """
 @recipe Tooltip begin
     # General
@@ -52,6 +54,15 @@ Creates a tooltip pointing at `position` displaying the given `string
 
     mixin_generic_plot_attributes()...
     inspectable = false
+end
+
+function attribute_groups(::Type{<:Tooltip})
+    groups = default_attribute_groups()
+    attr = uncategorized_attributes(Text)
+    filter!(!=(:offset), attr)
+    push!(attr, :textcolor)
+    push!(groups, "Text" => attr)
+    return groups
 end
 
 function convert_arguments(::Type{<:Tooltip}, x::Real, y::Real, str::AbstractString)