+++ Makie/src/basic_recipes/volumeslices.jl
@@ -1,7 +1,10 @@
 """
-    volumeslices(x, y, z, v)
+Draws heatmap slices visualizing an xy, yz and xz plane of the volume data v.
 
-Draws heatmap slices of the volume `v`.
+## Arguments
+
+* `x, y, z, v` Defines the extends of the volume with `x, y, z` (any type that implements `extrema`)
+    and the volume data `v` (an `AbstractArray{3, Real}`).
 """
 @recipe VolumeSlices (x, y, z, volume) begin
     documented_attributes(Heatmap)...
@@ -11,7 +14,7 @@ Draws heatmap slices of the volume `v`.
     bbox_color = RGBAf(0.5, 0.5, 0.5, 0.5)
 end
 
-function Makie.plot!(plot::VolumeSlices)
+function plot!(plot::VolumeSlices)
     @extract plot (x, y, z, volume)
 
     map!(plot.attributes, [:colorrange, :volume], :computed_colorrange) do colorrange, volume