+++ Makie/src/basic_recipes/waterfall.jl
@@ -1,6 +1,4 @@
 """
-    waterfall(x, y; kwargs...)
-
 Plots a [waterfall chart](https://en.wikipedia.org/wiki/Waterfall_chart) to visualize individual
 positive and negative components that add up to a net result as a barplot with stacked bars next
 to each other.
@@ -30,9 +28,28 @@ to each other.
     final_dodge_gap = 0
 end
 
-conversion_trait(::Type{<:Waterfall}) = PointBased()
+function attribute_groups(::Type{<:Waterfall})
+    groups = attribute_groups(BarPlot)
+    push!(
+        groups, "Marker Attributes" => sort!(
+            [
+                :show_direction, :marker_pos, :marker_neg, :direction_color,
+            ]
+        )
+    )
+    push!(
+        groups, "Final Bar Attributes" => sort!(
+            [
+                :show_final, :final_color, :final_gap, :final_dodge_gap,
+            ]
+        )
+    )
+    return groups
+end
 
-function Makie.plot!(p::Waterfall)
+conversion_trait(::Type{<:Waterfall}) = PointBased2D()
+
+function plot!(p::Waterfall)
     function stack_bars(xy, dodge, stack)
         x, y = first.(xy), last.(xy)
         if stack === automatic