+++ Makie/src/documentation/docstringextension.jl
@@ -6,6 +6,8 @@
 #                        Attributes                        #
 ############################################################
 
+# This only exists for Axis3D/axis3d!() (aka OldAxis) now...
+
 struct DocThemer <: DocStringExtensions.Abbreviation end
 
 const ATTRIBUTES = DocThemer()
@@ -18,6 +20,9 @@ end
 ############################################################
 #                        Instances                         #
 ############################################################
+
+# This allows you to add `$INSTANCES` in a docstring of an enum to splice in
+# a table of the enum names and values
 
 struct DocInstances <: DocStringExtensions.Abbreviation end
 
@@ -43,21 +48,3 @@ function DocStringExtensions.format(::DocInstances, buf, doc)
     # print the Markdown table into the buffer
     return show(buf, Markdown.MD(Markdown.Table(rows, [:l, :l])))
 end
-
-# """
-# A lolly instead of a lol!
-#
-# $(INSTANCES)
-# """
-# @enum Lolly Pop Bang Snap Crackle Jerk
-#
-# @doc Lolly
-# A lolly instead of a lol!
-#
-# Instance Value
-# –––––––– –––––
-# Pop      0
-# Bang     1
-# Snap     2
-# Crackle  3
-# Jerk     4